Fix OpenSuggestionCommandTest
